PLANET-7611: Rename Happy Point
- The block name Happy Point is now used consistently
