+++ assets/src/scss/blocks/HappyPoint/HappyPointEditorStyle.scss
