+++ assets/src/scss/blocks/HappyPoint/HappyPointStyle.scss
