+++ assets/src/scss/editorStyle.scss
@@ -41,7 +41,7 @@
 @import "blocks/Columns/ColumnsEditorStyle";
 @import "blocks/Cookies/CookiesEditorStyle";
 @import "blocks/ENForm/ENFormEditorStyle";
-@import "blocks/HappyPoint/HappypointEditorStyle";
+@import "blocks/HappyPoint/HappyPointEditorStyle";
 @import "blocks/TableOfContents/TableOfContentsEditorStyle";
 @import "blocks/TakeActionBoxout/TakeActionBoxoutEditorStyle";
 @import "blocks/Timeline/TimelineEditorStyle";
